+++ src/arrow-switcher/api/zh-cn.js
@@ -45,6 +45,18 @@ module.exports = [
                 description: '点击下一条事件',
                 type: 'EventEmitter<ThyArrowSwitcherEvent>',
                 default: '$event: { index: number, event: Event}'
+            },
+            {
+                name: 'thyPreviousTooltip',
+                description: '设置上一条 Hover Tooltip 提示',
+                type: 'string',
+                default: '-'
+            },
+            {
+                name: 'thyNextTooltip',
+                description: '设置下一条 Hover Tooltip 提示',
+                type: 'string',
+                default: '-'
             }
         ]
     }