+++ src/core/event-dispatchers/index.ts
@@ -1,3 +1,1 @@
-export * from './click-dispatcher';
-export * from './keyboard-dispatcher';
-export * from './event-dispatcher';
+export * from '@tethys/cdk/event';