+++ src/skeleton/skeleton-rectangle.component.ts
@@ -39,7 +39,6 @@ export class ThySkeletonRectangleComponent implements OnInit, OnChanges {
      * 是否开启动画
      */
     @Input()
-    @InputBoolean()
     thyAnimated: boolean;
 
     /**
@@ -115,7 +114,9 @@ export class ThySkeletonRectangleComponent implements OnInit, OnChanges {
                     0.15
                 )}, ${helpers.hexToRgb(this.thySecondaryColor, 0)}`
             }),
-            animation: this.thyAnimated !== false ? `thy-skeleton-animation ${this.thyAnimatedInterval}s infinite` : 'none'
+            animation: ![false, 'false'].includes(this.thyAnimated)
+                ? `thy-skeleton-animation ${this.thyAnimatedInterval}s infinite`
+                : 'none'
         };
     }