+++ src/table/index.ts
@@ -4,3 +4,4 @@ export * from './pipes/table.pipe';
 export * from './pipes/drag.pipe';
 export * from './table.module';
 export * from './table-column.component';
+export * from './table-skeleton.component';